refactor: organize agent run service modules
