+++ packages/server/src/agents/run/queue.ts
@@ -1,6 +1,6 @@
-import type { AgentOutputRow, AgentOutputTriggerType } from "../db/repositories/agent-outputs";
-import type { Logger } from "../logger";
-import type { QueueManager } from "../queue";
+import type { AgentOutputRow, AgentOutputTriggerType } from "../../db/repositories/agent-outputs";
+import type { Logger } from "../../logger";
+import type { QueueManager } from "../../queue";
 
 export interface ScheduledRunAdmission {
   controller: AbortController;
